requirements.txt
pandas
numpy
matplotlib.pyplot
matplotlib
plotly.express